Fixed traits.
