+++ include/nnlib/nn/identity.h
@@ -30,7 +30,7 @@ public:
 	}
 	
 	// MARK: Serialization
-	
+	/*
 	/// \brief Write to an archive.
 	///
 	/// \param out The archive to which to write.
@@ -54,6 +54,7 @@ public:
 		in >> shape;
 		this->inputs(shape);
 	}
+	*/
 };
 
 NNSerializable(Identity<double>, Module<double>);
