+++ include/nnlib/nn/relu.h
@@ -48,7 +48,7 @@ public:
 	}
 	
 	// MARK: Serialization
-	
+	/*
 	/// \brief Write to an archive.
 	///
 	/// \param out The archive to which to write.
@@ -72,7 +72,7 @@ public:
 		in >> shape;
 		this->inputs(shape);
 	}
-	
+	*/
 private:
 	T m_leak;
 };
