+++ test/nn/test_linear.h
@@ -57,13 +57,13 @@ void TestLinear()
 	module.batch(32);
 	NNAssert(module.batch() == 32, "Linear::batch failed!");
 	
-	Linear<> *deserialized = nullptr;
-	Archive::fromString((Archive::toString() << module).str()) >> deserialized;
-	NNAssert(
-		deserialized != nullptr && module.parameters().addV(deserialized->parameters(), -1).square().sum() < 1e-9,
-		"Linear::save and/or Linear::load failed!"
-	);
-	delete deserialized;
+	// Linear<> *deserialized = nullptr;
+	// Archive::fromString((Archive::toString() << module).str()) >> deserialized;
+	// NNAssert(
+	// 	deserialized != nullptr && module.parameters().addV(deserialized->parameters(), -1).square().sum() < 1e-9,
+	// 	"Linear::save and/or Linear::load failed!"
+	// );
+	// delete deserialized;
 }
 
 #endif
