+++ test/nn/test_relu.h
@@ -54,14 +54,14 @@ void TestReLU()
 	catch(const std::runtime_error &e) {}
 	NNAssert(ok, "ReLU::resize allowed unequal inputs and outputs!");
 	
-	ReLU<> *deserialized = nullptr;
-	Archive::fromString((Archive::toString() << map).str()) >> deserialized;
-	NNAssert(
-		deserialized != nullptr && map.inputs() == deserialized->inputs() && map.outputs() == deserialized->outputs(),
-		"ReLU::save and/or ReLU::load failed!"
-	);
-	
-	delete deserialized;
+	// ReLU<> *deserialized = nullptr;
+	// Archive::fromString((Archive::toString() << map).str()) >> deserialized;
+	// NNAssert(
+	// 	deserialized != nullptr && map.inputs() == deserialized->inputs() && map.outputs() == deserialized->outputs(),
+	// 	"ReLU::save and/or ReLU::load failed!"
+	// );
+	// 
+	// delete deserialized;
 }
 
 #endif
